# docs/_static/openapi.yaml
openapi: 3.0.3
info:
  title: ''
  version: 0.0.0
paths:
...
  /api/dataset-distinct-values/{dataset_id}/{field}/:
    get:
      operationId: dataset_distinct_values
      description: Return the field's 200 most common values, with their counts, in
        descending order.
      parameters:
      - in: path
        name: dataset_id
        schema:
          type: integer
        required: true
      - in: path
        name: field
        schema:
          type: string
        required: true
      tags:
      - dataset-distinct-values
      security:
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DistinctValue'
          description: ''
  /api/dataset-distinct-values/{dataset_id}/{field}/{query}/:
    get:
      operationId: dataset_distinct_values_search
      description: Return the field's 200 most common values containing the query,
        with their counts, in descending order.
      parameters:
      - in: path
        name: dataset_id
        schema:
          type: integer
        required: true
      - in: path
        name: field
        schema:
          type: string
        required: true
      - in: path
        name: query
        schema:
          type: string
        required: true
      tags:
      - dataset-distinct-values
      security:
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DistinctValue'
          description: ''
...
components:
  schemas:
...
    DistinctValue:
      type: object
      properties:
        value:
          type: string
          description: The field's value
        count:
          type: integer
          description: The number of data items with this value
      required:
      - count
      - value
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic